+++ src/main/java/com/kingpixel/ultrasts/commands/base/ReloadCommand.java
@@ -18,7 +18,7 @@ public class ReloadCommand {
             UltraSTS.reload();
             context.getSource().sendMessage(
               Text.literal(
-                "UltraSTS reloaded successfully!"
+                UltraSTS.lang.getReloadSuccess()
               )
             );
             return 1;